feat: always render label tooltip in the DOM
