+++ packages/ui/src/components/Tooltip/TooltipElement.tsx
@@ -1,6 +1,7 @@
 import { cn } from '@ultraviolet/utils'
 import type { TooltipProps } from '.'
 import type { useTooltip } from './useTooltip'
+import { visuallyHiddenStyle } from '../VisuallyHidden/styles.css'
 import { tooltipStyle } from './styles.css'
 
 type TooltipElementProps = Pick<TooltipProps, 'id' | 'className' | 'style' | 'data-testid' | 'maxWidth' | 'text'> & {
@@ -23,7 +24,8 @@ export function TooltipElement({
       className={cn(
         className,
         tooltipStyle.tooltip,
-        tooltip.status === 'close' ? tooltipStyle.animation.exit : tooltipStyle.animation.enter,
+        tooltip.isMounted ? null : visuallyHiddenStyle.visuallyHidden,
+        tooltip.status === 'open' || tooltip.status === 'close' ? tooltipStyle.animation[tooltip.status] : null,
       )}
       data-placement={tooltip.finalPlacement}
       data-testid={dataTestId}
